refactor: gql 관련 파일 Apollo 폴더로 이동
