+++ frontend/src/App.tsx
@@ -7,15 +7,15 @@ import { userIdVar } from '.';
 import Admin from './components/Admin';
 import ChannelList from './components/ChannelList';
 import LoadingBackdrop from './components/commons/LoadingBackdrop';
+import RestrictRoute from './components/commons/RestrictRoute';
 import Home from './components/Home';
 import Login from './components/Login';
 import MyProfile from './components/MyProfile';
 import Profile from './components/Profile';
 import Rank from './components/Rank';
 import UserRank from './components/UserRank';
-import { GET_MY_ID } from './utils/gqls';
-import { GetMyIdResponse } from './utils/responseModels';
-import RestrictRoute from './utils/RestrictRoute';
+import { GET_MY_ID } from './utils/Apollo/gqls';
+import { GetMyIdResponse } from './utils/Apollo/responseModels';
 
 function App(): JSX.Element {
   const { loading, data, error } = useQuery<GetMyIdResponse>(GET_MY_ID);
